fix: UI and add Nprogress Bar
