+++ src/main.js
@@ -4,7 +4,7 @@ import "./registerServiceWorker";
 import router from "./router";
 import store from "./store";
 import vuetify from "./plugins/vuetify";
-
+import "nprogress/nprogress.css";
 Vue.config.productionTip = false;
 
 new Vue({
