+++ src/router/index.js
@@ -2,6 +2,7 @@ import Vue from "vue";
 import VueRouter from "vue-router";
 import Home from "../views/Home.vue";
 import Signin from "../views/Signin.vue";
+
 Vue.use(VueRouter);
 
 const routes = [
